Improve cost calculation performance
