+++ src/components/Footer.tsx
@@ -73,4 +73,4 @@ const Footer: React.FC = () => {
   );
 };
 
-export default Footer;
+export default React.memo(Footer);
