+++ src/components/Header.tsx
@@ -201,4 +201,4 @@ const Header: React.FC<HeaderProps> = ({
   );
 };
 
-export default Header;
+export default React.memo(Header);
