+++ src/hooks/useCalculation.ts
@@ -1,96 +1,7 @@
-import { useState, useCallback } from 'react';
+import { useState, useCallback, useEffect, useRef } from 'react';
+import type { FormData, CalculationResults } from '@/utils/calc';
+import { calculateResults } from '@/utils/calc';
 
-export interface Worker {
-  id: string;
-  hourlyRate: number;
-  hours: number;
-}
-
-export interface ProcessingPhase {
-  id: string;
-  name: string;
-  wastePercentage: number;
-  addedWeight: number; // can be negative for loss, positive for glazing
-  description: string;
-}
-
-export interface FormData {
-  productName: string;
-  productType: 'fish' | 'squid' | 'octopus' | 'other';
-  purchasePrice: number;
-  quantity: number;
-  waste: number;
-  glazingPercent: number;
-  vatPercent: number;
-  workers: Worker[];
-  boxCost: number;
-  bagCost: number;
-  distance: number;
-  fuelCost: number;
-  tolls: number;
-  parkingCost: number;
-  driverSalary: number;
-  profitMargin: number;
-  profitTarget: number;
-  competitor1: number;
-  competitor2: number;
-  electricityCost: number;
-  equipmentCost: number;
-  insuranceCost: number;
-  rentCost: number;
-  communicationCost: number;
-  otherCosts: number;
-  originAddress: string;
-  destinationAddress: string;
-  routeCalculated: boolean;
-  estimatedDuration: string;
-  // Premium fields
-  batchNumber: string;
-  supplierName: string;
-  processingPhases: ProcessingPhase[];
-  targetSellingPrice: number;
-  minimumMargin: number;
-  storageTemperature: number;
-  shelfLife: number;
-  certifications: string[];
-  customerPrice: number;
-  seasonalMultiplier: number;
-}
-
-export interface CalculationResults {
-  totalCost: number;
-  totalCostWithVat: number;
-  sellingPrice: number;
-  profitPerKg: number;
-  profitMargin: number;
-  netWeight: number;
-  purchaseCost: number;
-  laborCost: number;
-  packagingCost: number;
-  transportCost: number;
-  additionalCosts: number;
-  vatAmount: number;
-  // Premium results
-  finalProcessedWeight: number;
-  totalWastePercentage: number;
-  costBreakdown: Array<{
-    category: string;
-    amount: number;
-    percentage: number;
-  }>;
-  recommendedSellingPrice: number;
-  competitorAnalysis: {
-    ourPrice: number;
-    competitor1Diff: number;
-    competitor2Diff: number;
-    marketPosition: 'competitive' | 'expensive' | 'cheap';
-  };
-  profitAnalysis: {
-    breakEvenPrice: number;
-    marginAtCurrentPrice: number;
-    recommendedMargin: number;
-  };
-}
 
 export const useCalculation = () => {
   const [formData, setFormData] = useState<Partial<FormData>>({
@@ -146,152 +57,36 @@ export const useCalculation = () => {
     setFormData(prev => ({ ...prev, ...updates }));
   }, []);
 
-  const calculatePhaseResult = (inputWeight: number, lossPct: number, addPct: number): number => {
-    // Apply loss first
-    const afterLoss = inputWeight * (1 - lossPct / 100);
-    // Then apply addition based on the weight after loss
-    const outputWeight = afterLoss + afterLoss * (addPct / 100);
-    return outputWeight;
-  };
+  const workerRef = useRef<Worker>();
+
+  useEffect(() => {
+    if (typeof Worker !== 'undefined') {
+      workerRef.current = new Worker(new URL('../workers/calculateWorker.ts', import.meta.url), { type: 'module' });
+    }
+    return () => {
+      workerRef.current?.terminate();
+    };
+  }, []);
 
   const calculate = useCallback(async (): Promise<void> => {
     setIsCalculating(true);
-    
-    // Simulate calculation delay for better UX
-    await new Promise(resolve => setTimeout(resolve, 800));
 
-    try {
-      // Advanced processing calculation with multiple phases - Serial calculations
-      let currentWeight = formData.quantity || 0;
-      let totalWastePercentage = 0;
-      
-      // Apply processing phases sequentially
-      if (formData.processingPhases && formData.processingPhases.length > 0) {
-        formData.processingPhases.forEach((phase, index) => {
-          console.log(`Phase ${index + 1}: ${phase.name}`);
-          console.log(`Input weight: ${currentWeight.toFixed(2)} kg`);
-          
-          if (phase.wastePercentage > 0 || phase.addedWeight !== 0) {
-            // Use the updated calculatePhaseResult function
-            const newWeight = calculatePhaseResult(currentWeight, phase.wastePercentage, phase.addedWeight);
-            
-            console.log(`After ${phase.name}: ${newWeight.toFixed(2)} kg`);
-            console.log(`Loss: ${phase.wastePercentage}%, Added: ${phase.addedWeight}%`);
-            
-            // Track total waste percentage (cumulative effect)
-            if (phase.wastePercentage > 0) {
-              totalWastePercentage += phase.wastePercentage;
-            }
-            
-            currentWeight = newWeight;
-          }
-        });
-      } else {
-        // Fallback to simple calculation
-        const netWeight = currentWeight * (1 - (formData.waste || 0) / 100);
-        currentWeight = netWeight * (1 + (formData.glazingPercent || 0) / 100);
-        totalWastePercentage = formData.waste || 0;
-      }
-
-      const finalProcessedWeight = currentWeight;
-      
-      console.log(`Final processed weight: ${finalProcessedWeight.toFixed(2)} kg`);
-      console.log(`Total waste percentage: ${totalWastePercentage.toFixed(1)}%`);
-      
-      const purchaseCost = (formData.purchasePrice || 0) * (formData.quantity || 0);
-      
-      // Calculate total labor cost from all workers
-      const laborCost = (formData.workers || []).reduce((sum, worker) => 
-        sum + (worker.hourlyRate * worker.hours), 0
-      );
-      
-      const packagingCost = (formData.boxCost || 0) + (formData.bagCost || 0);
-      const transportCost = 
-        (formData.distance || 0) * (formData.fuelCost || 0) + 
-        (formData.tolls || 0) + 
-        (formData.parkingCost || 0) +
-        (formData.driverSalary || 0);
-      
-      const additionalCosts = 
-        (formData.electricityCost || 0) + 
-        (formData.equipmentCost || 0) + 
-        (formData.insuranceCost || 0) + 
-        (formData.rentCost || 0) + 
-        (formData.communicationCost || 0) + 
-        (formData.otherCosts || 0);
-
-      const totalCost = purchaseCost + laborCost + packagingCost + transportCost + additionalCosts;
-      
-      // Calculate VAT
-      const vatAmount = totalCost * ((formData.vatPercent || 0) / 100);
-      const totalCostWithVat = totalCost + vatAmount;
-      
-      // Apply seasonal multiplier
-      const seasonalAdjustment = (formData.seasonalMultiplier || 1);
-      const adjustedCost = totalCostWithVat * seasonalAdjustment;
-      
-      const sellingPrice = adjustedCost * (1 + (formData.profitMargin || 0) / 100);
-      const sellingPricePerKg = sellingPrice / Math.max(finalProcessedWeight, 0.001);
-      
-      const profitPerKg = sellingPricePerKg - (adjustedCost / Math.max(finalProcessedWeight, 0.001));
-
-      // Cost breakdown for premium analysis
-      const costBreakdown = [
-        { category: 'Αγορά', amount: purchaseCost, percentage: (purchaseCost / totalCost) * 100 },
-        { category: 'Εργασία', amount: laborCost, percentage: (laborCost / totalCost) * 100 },
-        { category: 'Συσκευασία', amount: packagingCost, percentage: (packagingCost / totalCost) * 100 },
-        { category: 'Μεταφορά', amount: transportCost, percentage: (transportCost / totalCost) * 100 },
-        { category: 'Λοιπά', amount: additionalCosts, percentage: (additionalCosts / totalCost) * 100 }
-      ].filter(item => item.amount > 0);
-
-      // Competitor analysis
-      const competitor1Diff = (formData.competitor1 || 0) - sellingPricePerKg;
-      const competitor2Diff = (formData.competitor2 || 0) - sellingPricePerKg;
-      let marketPosition: 'competitive' | 'expensive' | 'cheap' = 'competitive';
-      
-      if (competitor1Diff > 0.5 || competitor2Diff > 0.5) {
-        marketPosition = 'cheap';
-      } else if (competitor1Diff < -0.5 || competitor2Diff < -0.5) {
-        marketPosition = 'expensive';
-      }
-
-      // Profit analysis
-      const breakEvenPrice = adjustedCost / Math.max(finalProcessedWeight, 0.001);
-      const marginAtCurrentPrice = ((sellingPricePerKg - breakEvenPrice) / sellingPricePerKg) * 100;
-      const recommendedMargin = Math.max(formData.minimumMargin || 15, 20);
-      
-      // Recommended selling price based on market analysis
-      const recommendedSellingPrice = breakEvenPrice * (1 + recommendedMargin / 100);
-
-      setResults({
-        totalCost,
-        totalCostWithVat: adjustedCost,
-        sellingPrice: sellingPricePerKg,
-        profitPerKg,
-        profitMargin: formData.profitMargin || 0,
-        netWeight: finalProcessedWeight,
-        purchaseCost,
-        laborCost,
-        packagingCost,
-        transportCost,
-        additionalCosts,
-        vatAmount,
-        finalProcessedWeight,
-        totalWastePercentage,
-        costBreakdown,
-        recommendedSellingPrice,
-        competitorAnalysis: {
-          ourPrice: sellingPricePerKg,
-          competitor1Diff,
-          competitor2Diff,
-          marketPosition
-        },
-        profitAnalysis: {
-          breakEvenPrice,
-          marginAtCurrentPrice,
-          recommendedMargin
+    const run = () => {
+      return new Promise<CalculationResults>((resolve) => {
+        if (workerRef.current) {
+          workerRef.current.onmessage = (e: MessageEvent<CalculationResults>) => {
+            resolve(e.data);
+          };
+          workerRef.current.postMessage(formData as FormData);
+        } else {
+          resolve(calculateResults(formData as FormData));
         }
       });
+    };
+
+    try {
+      const result = await run();
+      setResults(result);
     } catch (error) {
       console.error('Calculation error:', error);
       setResults(null);
